move socket.io/admin.settings to its own file
